feat: update app manifest to use agent_view
Flips both Casey implementations from assistant_view to agent_view in
preparation for the new Agent DM Messages Tab experience:

- Rename features.assistant_view -> features.agent_view in both
  manifests, and the inner assistant_description -> agent_description.
- Remove assistant_thread_started from event subscriptions.
- Merge setSuggestedPrompts into app_home_opened, gated on event.tab
  so the Home tab keeps publishing its Block Kit view (category
  buttons + MCP status) and the Messages tab pins the suggested
  prompts.
- Delete the now-unused assistant-thread-started.js handlers and
  their tests, and refresh app-home-opened.test.js to cover both
  event.tab branches.

TODO(agent-dm-messages-tab) markers note where app_context_changed
wiring will land once Bolt JS exposes the event type.
