+++ claude-agent-sdk/listeners/events/index.js
@@ -1,6 +1,5 @@
 import { handleAppHomeOpened } from './app-home-opened.js';
 import { handleAppMentioned } from './app-mentioned.js';
-import { handleAssistantThreadStarted } from './assistant-thread-started.js';
 import { handleMessage } from './message.js';
 
 /**
@@ -11,6 +10,5 @@ import { handleMessage } from './message.js';
 export function register(app) {
   app.event('app_home_opened', handleAppHomeOpened);
   app.event('app_mention', handleAppMentioned);
-  app.event('assistant_thread_started', handleAssistantThreadStarted);
   app.event('message', handleMessage);
 }
